requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
httpx
google-api-python-client==2.139.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
apscheduler==3.10.4
sqlalchemy==2.0.35
aiosqlite==0.20.0
google-genai==1.10.0
psycopg2-binary==2.9.9
python-jose[cryptography]
passlib[bcrypt]
python-multipart